Allow starting or stopping bundles. Basically, – if the bundle is `Active` – assume we're stopping it and, if otherwise, assume we're attempting to start it.
I figured it'd make for quicker usage for users since you don't have
to remember which you want to do (plus, trying to start an `Active`
bundle doesn't actually do anything so simplify UX by just stopping
the bundle and vice versa).

It also allows for applying to several bundles, at once. Since the
`revert-buffer` call has to make a `url` call (which sends a
`message`), just display the state of the last buffer.

Maybe, though, I'll just store it in a list and display it as one
`message` at the end…. I can do that for the next commit.