This is a mess of parsing since it's frickin' HTML that it returns, like a loser, so things may break if they change anything any but it works, for now.
I'm not sure how I feel on the behavior (and may change it to allow
for user adjustment) but I think it makes sense: if you're not in the
list or info. buffer, just assume the user will get back to it and
call the function `get-users` (maybe I'll add a check, later, for if a
buffer exists and refresh that one…) and that they don't want to
change their current buffer, any; if you're in a list buffer, refresh
it to get the new user; and, if you're in the info. buffer, switch it
to the new user's profile so they can view the new user (worst case,
they can hit `back` and get to the previous user they were viewing).