Add the ability to open the bundle in the browser.
It seems like you can view more details of individual bundles by
specifying them by their ID with ".json" at the end but no way to do
it from the "/system/console/bundles.json" endpoint (because of
course).

I may add a command to ping the former URL and see if there's any
errors and then display it as a `message`; or, Hell, maybe retrieve
the data, when requested, and open it in a new Info. buffer.

Not sure if there's much purpose for the latter, though, if you can
just open the bundle directly in the browser, if we're already making
an HTTP request for the extra info. We'll see.

Either way, no point in making however many HTTP calls for every
bundle just to get extra data.