Turns out AEM has a _third_ endpoint that we can use to get package details (/crx/packmgr/list.jsp) which doesn't require parsing through subnodes.
It even seems to have all of the info I pulled from the various nodes
of the packages (well, nearly; it doesn't seem to have UUID, for
example, but I can certainly live with that. Probably everything in
that `vlt:definition` node _plus_ `:jcr:data` (as `size`) and
`jcr:path` (as `path`); only thing possibly missing seems to be
`jcr:uuid` so it should be an easy switch.

I don't want to just obliterate the previous work so separating this
out into, yet another, domain name for functions ("extensive").

And then we'll add the new endpoint and probably phase out the
previous two since the new one should cover everything I'd
expect/want.

At least the call should be faster, now, with less parsing; 'just need
to figure out that damn status….