Everything's _pretty_ much the same, as before just added the `pid` and switched `:jcr:data` to `size`, along with removing the `jcr:` from a few of them. Pretty much like I said, originally.
The only real difference is I spotted a `groupTitle` property. I dunno
if it was present for the other one but adding it, just in case. Even
if I eventually decide not to use it, keeping it in the `:titles` is
useful and an easy way to keep track of the fact that it exists.